requirements-macos.txt
Flask==3.1.0
pywebview==6.0
Jinja2==3.1.5
click==8.1.8
Werkzeug==3.1.3
itsdangerous==2.2.0
MarkupSafe==3.0.2
requests==2.32.5
urllib3==2.5.0
certifi==2025.1.31
charset-normalizer==3.4.3
idna==3.10
httpcore==1.0.7
httpx==0.28.1
h11==0.14.0
pydantic==2.10.6
pydantic_core==2.27.2
pillow==11.1.0
numpy==2.3.2
psutil==6.1.1
cloudinary==1.44.1
# macOS 专用依赖
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-WebKit